Add initial `tarjan_scc` from tests to algo
